# e-commerce-project/public/swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "E-commerce-API",
    "description": "Documentation technique de notre API pour notre projet E-commerce (Epitech)",
    "version": "0.1.9"
  },
  "servers": [
    {
      "url": "http://localhost:8080/",
      "description": "URL principal de l'API"
    },
    {
      "url": "http://localhost:3000/",
      "description": "URL du front-end pour tester le fonctionnement de l'API"
    }
  ],
  "paths": {
    "/api/register": {
      "post": {
        "tags": [
          "Utilisateurs"
        ],
        "summary": "Inscription de l'utilisateur",
        "description": "Cette méthode permet à l'utilisateur de créer son propre compte",
        "operationId": "createUser",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "login": {
                    "type": "string",
                    "example": "ugo-bertrand"
                  },
                  "password": {
                    "type": "string",
                    "example": "bipbip"
                  },
                  "email": {
                    "type": "string",
                    "example": "ugo.bertrand@epitech.eu"
                  },
                  "firstname": {
                    "type": "string",
                    "example": "ugo"
                  },
                  "lastname": {
                    "type": "string",
                    "example": "bertrand"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "L'utilisateur a été créer"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Le contenu ne doit pas être vide, veuillez mettre les valeurs correspondantes"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Internal Server Error"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/login": {
      "get": {
        "tags": [
          "Utilisateurs"
        ],
        "summary": "Connexion de l'utilsateur",
        "description": "Cette méthode permet à l'utilisateur de se connecter à son compte",
        "operationId": "loginUser",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "login": {
                    "type": "string",
                    "example": "ugo-bertrand"
                  },
                  "password": {
                    "type": "string",
                    "example": "bipbip"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string",
                      "example": "eyakljgsfsdfesgjsmlsqkjfùsmqlkhsdmlfsdfshdfsdfsd"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Le mot de passe ne correspond pas"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Le contenu ne doit pas être vide, veuillez mettre les valeurs correspondantes"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Internal Server Error"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/users": {
      "put": {
        "tags": [
          "Utilisateurs"
        ],
        "summary": "Mettre à jour les données de l'utilisateur",
        "description": "Cette méthode permet à l'utilisateur de modifier son compte",
        "operationId": "updateUser",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "type": "string",
            "required": true,
            "description": "Token pour vérifier si l'utilisateur est bien connecté"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "login": {
                    "type": "string",
                    "example": "ugo-bertrand"
                  },
                  "password": {
                    "type": "string",
                    "example": "bipbip"
                  },
                  "email": {
                    "type": "string",
                    "example": "ugo.bertrand@epitech.eu"
                  },
                  "firstname": {
                    "type": "string",
                    "example": "ugo"
                  },
                  "lastname": {
                    "type": "string",
                    "example": "bertrand"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statut": {
                      "type": "string",
                      "example": "Votre utilisateur a bien été modifié"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Vous n'êtes pas connecté, veuillez vous connecter"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Le contenu ne doit pas être vide, veuillez mettre les valeurs correspondantes"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Internal Server Error"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/users/": {
      "get": {
        "tags": [
          "Utilisateurs"
        ],
        "summary": "Récupération des informations de l'utilisateurs",
        "description": "Cette méthode permet à l'utilisateur de récupérer ses informations",
        "operationId": "displayUser",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "Token pour vérifier si l'utilisateur est bien connecté"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "login": {
                      "type": "string",
                      "example": "ugo-bertrand"
                    },
                    "email": {
                      "type": "string",
                      "example": "ugo.bertrand@epitech.eu"
                    },
                    "firstname": {
                      "type": "string",
                      "example": "ugo"
                    },
                    "lastname": {
                      "type": "string",
                      "example": "bertrand"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Vous n'êtes pas connecté, veuillez vous connecter"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Internal Server Error"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}